LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/transform-translate-composition-expected.txt


FAIL Compositing: property <transform> underlying [translate(10px, 20px)] from add [translate(100px, 200px)] to add [translate(200px, 400px)] at (-0.5) should be [translate(60px, 120px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 60 , 120 ) " but got "matrix ( 1 , 0 , 0 , 1 , 50 , 100 ) "
FAIL Compositing: property <transform> underlying [translate(10px, 20px)] from add [translate(100px, 200px)] to add [translate(200px, 400px)] at (0) should be [translate(110px, 220px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 110 , 220 ) " but got "matrix ( 1 , 0 , 0 , 1 , 100 , 200 ) "
FAIL Compositing: property <transform> underlying [translate(10px, 20px)] from add [translate(100px, 200px)] to add [translate(200px, 400px)] at (0.25) should be [translate(135px, 270px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 135 , 270 ) " but got "matrix ( 1 , 0 , 0 , 1 , 125 , 250 ) "
FAIL Compositing: property <transform> underlying [translate(10px, 20px)] from add [translate(100px, 200px)] to add [translate(200px, 400px)] at (0.5) should be [translate(160px, 320px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 160 , 320 ) " but got "matrix ( 1 , 0 , 0 , 1 , 150 , 300 ) "
FAIL Compositing: property <transform> underlying [translate(10px, 20px)] from add [translate(100px, 200px)] to add [translate(200px, 400px)] at (0.75) should be [translate(185px, 370px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 185 , 370 ) " but got "matrix ( 1 , 0 , 0 , 1 , 175 , 350 ) "
FAIL Compositing: property <transform> underlying [translate(10px, 20px)] from add [translate(100px, 200px)] to add [translate(200px, 400px)] at (1) should be [translate(210px, 420px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 210 , 420 ) " but got "matrix ( 1 , 0 , 0 , 1 , 200 , 400 ) "
FAIL Compositing: property <transform> underlying [translate(10px, 20px)] from add [translate(100px, 200px)] to add [translate(200px, 400px)] at (1.5) should be [translate(260px, 520px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 260 , 520 ) " but got "matrix ( 1 , 0 , 0 , 1 , 250 , 500 ) "
FAIL Compositing: property <transform> underlying [translate(10px, 20px)] from add [translate(100px, 200px)] to replace [translate(210px, 420px)] at (-0.5) should be [translate(60px, 120px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 60 , 120 ) " but got "matrix ( 1 , 0 , 0 , 1 , 45 , 90 ) "
FAIL Compositing: property <transform> underlying [translate(10px, 20px)] from add [translate(100px, 200px)] to replace [translate(210px, 420px)] at (0) should be [translate(110px, 220px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 110 , 220 ) " but got "matrix ( 1 , 0 , 0 , 1 , 100 , 200 ) "
FAIL Compositing: property <transform> underlying [translate(10px, 20px)] from add [translate(100px, 200px)] to replace [translate(210px, 420px)] at (0.25) should be [translate(135px, 270px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 135 , 270 ) " but got "matrix ( 1 , 0 , 0 , 1 , 127.5 , 255 ) "
FAIL Compositing: property <transform> underlying [translate(10px, 20px)] from add [translate(100px, 200px)] to replace [translate(210px, 420px)] at (0.5) should be [translate(160px, 320px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 160 , 320 ) " but got "matrix ( 1 , 0 , 0 , 1 , 155 , 310 ) "
FAIL Compositing: property <transform> underlying [translate(10px, 20px)] from add [translate(100px, 200px)] to replace [translate(210px, 420px)] at (0.75) should be [translate(185px, 370px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 185 , 370 ) " but got "matrix ( 1 , 0 , 0 , 1 , 182.5 , 365 ) "
PASS Compositing: property <transform> underlying [translate(10px, 20px)] from add [translate(100px, 200px)] to replace [translate(210px, 420px)] at (1) should be [translate(210px, 420px)]
FAIL Compositing: property <transform> underlying [translate(10px, 20px)] from add [translate(100px, 200px)] to replace [translate(210px, 420px)] at (1.5) should be [translate(260px, 520px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 260 , 520 ) " but got "matrix ( 1 , 0 , 0 , 1 , 265 , 530 ) "
FAIL Compositing: property <transform> underlying [translateX(100px) rotate(90deg)] from add [translateX(100px)] to add [translateX(200px)] at (-0.5) should be [translateX(100px) rotate(90deg) translateX(50px)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 100 , 50 ) " but got "matrix ( 1 , 0 , 0 , 1 , 50 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px) rotate(90deg)] from add [translateX(100px)] to add [translateX(200px)] at (0) should be [translateX(100px) rotate(90deg) translateX(100px)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 100 , 100 ) " but got "matrix ( 1 , 0 , 0 , 1 , 100 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px) rotate(90deg)] from add [translateX(100px)] to add [translateX(200px)] at (0.25) should be [translateX(100px) rotate(90deg) translateX(125px)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 100 , 125 ) " but got "matrix ( 1 , 0 , 0 , 1 , 125 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px) rotate(90deg)] from add [translateX(100px)] to add [translateX(200px)] at (0.5) should be [translateX(100px) rotate(90deg) translateX(150px)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 100 , 150 ) " but got "matrix ( 1 , 0 , 0 , 1 , 150 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px) rotate(90deg)] from add [translateX(100px)] to add [translateX(200px)] at (0.75) should be [translateX(100px) rotate(90deg) translateX(175px)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 100 , 175 ) " but got "matrix ( 1 , 0 , 0 , 1 , 175 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px) rotate(90deg)] from add [translateX(100px)] to add [translateX(200px)] at (1) should be [translateX(100px) rotate(90deg) translateX(200px)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 100 , 200 ) " but got "matrix ( 1 , 0 , 0 , 1 , 200 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px) rotate(90deg)] from add [translateX(100px)] to add [translateX(200px)] at (1.5) should be [translateX(100px) rotate(90deg) translateX(250px)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 100 , 250 ) " but got "matrix ( 1 , 0 , 0 , 1 , 250 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px)] from accumulate [translateX(50px)] to accumulate [translateX(250px)] at (-0.5) should be [translateX(50px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 50 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , - 50 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px)] from accumulate [translateX(50px)] to accumulate [translateX(250px)] at (0) should be [translateX(150px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 150 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 50 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px)] from accumulate [translateX(50px)] to accumulate [translateX(250px)] at (0.25) should be [translateX(200px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 200 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 100 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px)] from accumulate [translateX(50px)] to accumulate [translateX(250px)] at (0.5) should be [translateX(250px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 250 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 150 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px)] from accumulate [translateX(50px)] to accumulate [translateX(250px)] at (0.75) should be [translateX(300px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 300 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 200 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px)] from accumulate [translateX(50px)] to accumulate [translateX(250px)] at (1) should be [translateX(350px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 350 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 250 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px)] from accumulate [translateX(50px)] to accumulate [translateX(250px)] at (1.5) should be [translateX(450px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 450 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 350 , 0 ) "
FAIL Compositing: property <transform> underlying [translateY(100px)] from accumulate [translateY(50px)] to accumulate [translateY(250px)] at (-0.5) should be [translateY(50px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 0 , 50 ) " but got "matrix ( 1 , 0 , 0 , 1 , 0 , - 50 ) "
FAIL Compositing: property <transform> underlying [translateY(100px)] from accumulate [translateY(50px)] to accumulate [translateY(250px)] at (0) should be [translateY(150px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 0 , 150 ) " but got "matrix ( 1 , 0 , 0 , 1 , 0 , 50 ) "
FAIL Compositing: property <transform> underlying [translateY(100px)] from accumulate [translateY(50px)] to accumulate [translateY(250px)] at (0.25) should be [translateY(200px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 0 , 200 ) " but got "matrix ( 1 , 0 , 0 , 1 , 0 , 100 ) "
FAIL Compositing: property <transform> underlying [translateY(100px)] from accumulate [translateY(50px)] to accumulate [translateY(250px)] at (0.5) should be [translateY(250px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 0 , 250 ) " but got "matrix ( 1 , 0 , 0 , 1 , 0 , 150 ) "
FAIL Compositing: property <transform> underlying [translateY(100px)] from accumulate [translateY(50px)] to accumulate [translateY(250px)] at (0.75) should be [translateY(300px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 0 , 300 ) " but got "matrix ( 1 , 0 , 0 , 1 , 0 , 200 ) "
FAIL Compositing: property <transform> underlying [translateY(100px)] from accumulate [translateY(50px)] to accumulate [translateY(250px)] at (1) should be [translateY(350px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 0 , 350 ) " but got "matrix ( 1 , 0 , 0 , 1 , 0 , 250 ) "
FAIL Compositing: property <transform> underlying [translateY(100px)] from accumulate [translateY(50px)] to accumulate [translateY(250px)] at (1.5) should be [translateY(450px)] assert_equals: expected "matrix ( 1 , 0 , 0 , 1 , 0 , 450 ) " but got "matrix ( 1 , 0 , 0 , 1 , 0 , 350 ) "
FAIL Compositing: property <transform> underlying [translateZ(100px)] from accumulate [translateZ(50px)] to accumulate [translateZ(250px)] at (-0.5) should be [translateZ(50px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 50 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , - 50 , 1 ) "
FAIL Compositing: property <transform> underlying [translateZ(100px)] from accumulate [translateZ(50px)] to accumulate [translateZ(250px)] at (0) should be [translateZ(150px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 150 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 50 , 1 ) "
FAIL Compositing: property <transform> underlying [translateZ(100px)] from accumulate [translateZ(50px)] to accumulate [translateZ(250px)] at (0.25) should be [translateZ(200px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 200 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 100 , 1 ) "
FAIL Compositing: property <transform> underlying [translateZ(100px)] from accumulate [translateZ(50px)] to accumulate [translateZ(250px)] at (0.5) should be [translateZ(250px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 250 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 150 , 1 ) "
FAIL Compositing: property <transform> underlying [translateZ(100px)] from accumulate [translateZ(50px)] to accumulate [translateZ(250px)] at (0.75) should be [translateZ(300px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 300 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 200 , 1 ) "
FAIL Compositing: property <transform> underlying [translateZ(100px)] from accumulate [translateZ(50px)] to accumulate [translateZ(250px)] at (1) should be [translateZ(350px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 350 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 250 , 1 ) "
FAIL Compositing: property <transform> underlying [translateZ(100px)] from accumulate [translateZ(50px)] to accumulate [translateZ(250px)] at (1.5) should be [translateZ(450px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 450 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 350 , 1 ) "
FAIL Compositing: property <transform> underlying [translate(100px, 50px)] from accumulate [translateZ(50px)] to accumulate [translateZ(250px)] at (-0.5) should be [translate3d(100px, 50px, -50px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 100 , 50 , - 50 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , - 50 , 1 ) "
FAIL Compositing: property <transform> underlying [translate(100px, 50px)] from accumulate [translateZ(50px)] to accumulate [translateZ(250px)] at (0) should be [translate3d(100px, 50px, 50px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 100 , 50 , 50 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 50 , 1 ) "
FAIL Compositing: property <transform> underlying [translate(100px, 50px)] from accumulate [translateZ(50px)] to accumulate [translateZ(250px)] at (0.25) should be [translate3d(100px, 50px, 100px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 100 , 50 , 100 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 100 , 1 ) "
FAIL Compositing: property <transform> underlying [translate(100px, 50px)] from accumulate [translateZ(50px)] to accumulate [translateZ(250px)] at (0.5) should be [translate3d(100px, 50px, 150px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 100 , 50 , 150 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 150 , 1 ) "
FAIL Compositing: property <transform> underlying [translate(100px, 50px)] from accumulate [translateZ(50px)] to accumulate [translateZ(250px)] at (0.75) should be [translate3d(100px, 50px, 200px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 100 , 50 , 200 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 200 , 1 ) "
FAIL Compositing: property <transform> underlying [translate(100px, 50px)] from accumulate [translateZ(50px)] to accumulate [translateZ(250px)] at (1) should be [translate3d(100px, 50px, 250px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 100 , 50 , 250 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 250 , 1 ) "
FAIL Compositing: property <transform> underlying [translate(100px, 50px)] from accumulate [translateZ(50px)] to accumulate [translateZ(250px)] at (1.5) should be [translate3d(100px, 50px, 350px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 100 , 50 , 350 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 350 , 1 ) "
FAIL Compositing: property <transform> underlying [translateX(100px) rotate(90deg)] from accumulate [translateX(100px)] to accumulate [translateX(200px)] at (-0.5) should be [translateX(150px) rotate(90deg)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 150 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 50 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px) rotate(90deg)] from accumulate [translateX(100px)] to accumulate [translateX(200px)] at (0) should be [translateX(200px) rotate(90deg)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 200 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 100 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px) rotate(90deg)] from accumulate [translateX(100px)] to accumulate [translateX(200px)] at (0.25) should be [translateX(225px) rotate(90deg)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 225 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 125 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px) rotate(90deg)] from accumulate [translateX(100px)] to accumulate [translateX(200px)] at (0.5) should be [translateX(250px) rotate(90deg)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 250 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 150 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px) rotate(90deg)] from accumulate [translateX(100px)] to accumulate [translateX(200px)] at (0.75) should be [translateX(275px) rotate(90deg)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 275 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 175 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px) rotate(90deg)] from accumulate [translateX(100px)] to accumulate [translateX(200px)] at (1) should be [translateX(300px) rotate(90deg)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 300 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 200 , 0 ) "
FAIL Compositing: property <transform> underlying [translateX(100px) rotate(90deg)] from accumulate [translateX(100px)] to accumulate [translateX(200px)] at (1.5) should be [translateX(350px) rotate(90deg)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 350 , 0 ) " but got "matrix ( 1 , 0 , 0 , 1 , 250 , 0 ) "
